# api/src/main/java/org/lfenergy/compas/scl/data/rest/PluginsCustomResourcesApi.java
// SPDX-FileCopyrightText: 2026 BearingPoint GmbH
//
// SPDX-License-Identifier: Apache-2.0
package org.lfenergy.compas.scl.data.rest;

import org.lfenergy.compas.scl.data.rest.api.plugins.resources.DataEntry;
import org.lfenergy.compas.scl.data.rest.api.plugins.resources.DataEntryWithContent;
import java.util.Date;

import org.lfenergy.compas.scl.data.rest.api.plugins.resources.PagedDataEntryResponse;
import java.util.UUID;
import org.lfenergy.compas.scl.data.rest.api.plugins.resources.UploadDataResponse;

import jakarta.ws.rs.*;


import java.io.InputStream;
import java.util.List;
import jakarta.validation.constraints.*;

/**
* Represents a collection of functions to interact with the API endpoints.
*/
@Path("/plugins/resources")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJAXRSSpecServerCodegen", comments = "Generator version: 7.12.0")
public interface PluginsCustomResourcesApi {

    /**
     * Delete all data entries for the given data type
     *
     * @param id Data type of the entries to delete
     * @return Data entries deleted successfully
     * @return No data entries found for the given data type
     * @return Internal server error
     */
    @DELETE
    @Path("/{id}")
    @Produces({ "application/json" })
    void deleteDataByType(@PathParam("id") String id);


    /**
     * Delete all data entries for the given data type and resource name
     *
     * @param dataType Data type of the entries to delete
     * @param name Name of the entries to delete
     * @return Data entries deleted successfully
     * @return No data entries found for the given data type and name
     * @return Internal server error
     */
    @DELETE
    @Path("/{data-type}/{name}")
    @Produces({ "application/json" })
    void deleteDataByTypeAndName(@PathParam("data-type") String dataType,@PathParam("name") String name);


    /**
     * Retrieve metadata for all uploaded data entries with optional filtering
     *
     * @param type Filter by data type
     * @param uploadedAfter Filter by upload date (ISO 8601 format)
     * @param uploadedBefore Filter by upload date (ISO 8601 format)
     * @param name Filter by name (contains search)
     * @param page Page number for pagination (0-indexed)
     * @param size Number of items per page
     * @return List of data entries retrieved successfully
     * @return Bad request - Invalid query parameters
     * @return Internal server error
     */
    @GET
    @Produces({ "application/json" })
    PagedDataEntryResponse getAllData(@QueryParam("type") @NotNull   String type,@QueryParam("uploadedAfter")   Date uploadedAfter,@QueryParam("uploadedBefore")   Date uploadedBefore,@QueryParam("name")   String name,@QueryParam("page") @Min(0) @DefaultValue("0")   Integer page,@QueryParam("size") @Min(1) @Max(100) @DefaultValue("20")   Integer size);


    /**
     * Returns all versions of the named resource as DataEntryWithContent objects
     *
     * @param dataType 
     * @param name 
     * @return List of versions with content
     * @return Resource not found
     * @return Internal server error
     */
    @GET
    @Path("/{data-type}/{name}/versions")
    @Produces({ "application/json" })
    List<DataEntryWithContent> getAllVersionsWithContentByTypeAndName(@PathParam("data-type") String dataType,@PathParam("name") String name);



    /**
     * Returns the specified version of the named resource, with content
     *
     * @param dataType
     * @param name
     * @param version
     * @return Resource version with content
     * @return Resource version not found
     * @return Internal server error
     */
    @GET
    @Path("/v1/{data-type}/{name}/{version}")
    @Produces({ "application/json" })
    DataEntryWithContent getSpecificVersionByTypeAndName(@PathParam("data-type") String dataType,@PathParam("name") String name,@PathParam("version") String version);

    /**
     * Retrieve a single data entry by its unique identifier, including the full content
     *
     * @param id Unique identifier of the data entry
     * @return Data entry retrieved successfully
     * @return Data entry not found
     * @return Internal server error
     */
    @GET
    @Path("/{id}")
    @Produces({ "application/json" })
    DataEntryWithContent getDataById(@PathParam("id") UUID id);


    /**
    * Retrieve all resources for the given data type, reduced to the latest version per resource name, without content
     *
    * @param dataType Data type of the entries to retrieve
    * @return Data entries retrieved successfully
    * @return Data entries not found
     * @return Internal server error
     */
    @GET
    @Path("/{data-type}/latest")
    @Produces({ "application/json" })
    List<DataEntry> getLatestDataByType(@PathParam("data-type") String dataType);


    /**
     * Retrieve the latest version of a data entry for the given data type and resource name, including the full content
     *
     * @param dataType Data type of the entry to retrieve
     * @param name Name of the entry to retrieve
     * @return Data entry retrieved successfully
     * @return Data entry not found
     * @return Internal server error
     */
    @GET
    @Path("/{data-type}/{name}/latest")
    @Produces({ "application/json" })
    DataEntryWithContent getLatestDataByTypeAndName(@PathParam("data-type") String dataType,@PathParam("name") String name);


    /**
     * Upload a JSON or XML data file with associated metadata
     *
     * @param type Type of the data being uploaded
     * @param name Name of the data file
     * @param contentType Content type of the uploaded file
     * @param content The JSON or XML file content
     * @param dataCompatibilityVersion Data compatibility version (semver format)
     * @param description Optional description of the data file
     * @param version Semantic version of the data file (semver format)
     * @param nextVersionType Type of the next version (used for automatic version increment) If provided, the service will automatically increment the version based on the specified type. If not provided, the version must be explicitly specified in the &#39;version&#39; field.
     * @return Data uploaded successfully
     * @return Bad request - Invalid input data
     * @return Conflict - Data with this name and version already exists
     * @return Payload too large
     * @return Unsupported media type
     * @return Internal server error
     */
    @POST
    @Consumes({ "multipart/form-data" })
    @Produces({ "application/json" })
    UploadDataResponse uploadData(@FormParam(value = "type")  String type,@FormParam(value = "name")  String name,@FormParam(value = "content-type")  String contentType, @FormParam(value = "content") InputStream content,@FormParam(value = "data-compatibility-version")  String dataCompatibilityVersion,@FormParam(value = "description")  String description,@FormParam(value = "version")  String version,@FormParam(value = "nextVersionType")  String nextVersionType);

}